Move handle_form_submission() into h.form
So it can be used by views outside of h.accounts.views
