+++ tests/h/form_test.py
@@ -5,6 +5,8 @@ import pytest
 
 from h import form
 
+from tests.h import conftest
+
 
 class TestJinja2Renderer(object):
 
@@ -111,3 +113,141 @@ class TestCreateForm(object):
     def pyramid_request(self, pyramid_request):
         pyramid_request.registry[form.ENVIRONMENT_KEY] = mock.sentinel.jinja2_env
         return pyramid_request
+
+
+class TestToXHRResponse(object):
+    """Unit tests for to_xhr_response()."""
+
+    def test_returns_given_result_if_not_xhr(self, pyramid_request):
+        """
+        If ``request`` isn't an XHR request it returns ``non_xhr_result``.
+
+        The calling view callable passes in the result that it would have
+        returned normally if this were not an XHR request as the
+        ``non_xhr_result`` argument. If the given ``request`` is not an XHR
+        request then ``non_xhr_result`` should just be returned unmodified.
+
+        """
+        pyramid_request.is_xhr = False
+        non_xhr_result = mock.Mock()
+
+        result = form.to_xhr_response(pyramid_request,
+                                      non_xhr_result,
+                                      mock.Mock())
+
+        assert result == non_xhr_result
+
+    def test_returns_form_if_xhr(self, pyramid_request):
+        """
+        If ``request`` is an XHR request it should return the rendered ``form``.
+
+        It should return ``form`` rendered to a ``<form>`` element HTML snippet.
+
+        """
+        pyramid_request.is_xhr = True
+        form_ = mock.Mock()
+
+        result = form.to_xhr_response(pyramid_request, mock.Mock(), form_)
+
+        assert result == form_.render.return_value
+
+
+@pytest.mark.usefixtures('to_xhr_response')
+class TestHandleFormSubmission(object):
+
+    def test_it_calls_validate(self, pyramid_request):
+        form_ = mock.Mock()
+
+        form.handle_form_submission(pyramid_request,
+                                    form_,
+                                    mock.Mock(),
+                                    mock.Mock())
+
+        form_.validate.assert_called_once_with(pyramid_request.POST.items())
+
+    def test_if_validation_fails_it_calls_on_failure(self,
+                                                     pyramid_request,
+                                                     invalid_form):
+        on_failure = mock.Mock()
+
+        form.handle_form_submission(pyramid_request,
+                                    invalid_form,
+                                    mock.Mock(),
+                                    on_failure)
+
+        on_failure.assert_called_once_with()
+
+    def test_if_validation_fails_it_calls_to_xhr_response(self,
+                                                          invalid_form,
+                                                          pyramid_request,
+                                                          to_xhr_response):
+        on_failure = mock.Mock()
+
+        form.handle_form_submission(pyramid_request,
+                                    invalid_form,
+                                    mock.Mock(),
+                                    on_failure)
+
+        to_xhr_response.assert_called_once_with(
+            pyramid_request, on_failure.return_value, invalid_form)
+
+    def test_if_validation_fails_it_returns_to_xhr_response(self,
+                                                            invalid_form,
+                                                            pyramid_request,
+                                                            to_xhr_response):
+        result = form.handle_form_submission(pyramid_request,
+                                             invalid_form,
+                                             mock.Mock(),
+                                             mock.Mock())
+
+        assert result == to_xhr_response.return_value
+
+    def test_if_validation_succeeds_it_calls_on_success(self, pyramid_request):
+        form_ = conftest.form_validating_to(mock.sentinel.appstruct)
+        on_success = mock.Mock()
+
+        form.handle_form_submission(pyramid_request,
+                                    form_,
+                                    on_success,
+                                    mock.Mock())
+
+        on_success.assert_called_once_with(mock.sentinel.appstruct)
+
+    def test_if_validation_succeeds_it_shows_a_flash_message(self,
+                                                             pyramid_request):
+        form.handle_form_submission(pyramid_request,
+                                    conftest.form_validating_to('anything'),
+                                    mock.Mock(),
+                                    mock.Mock())
+
+        assert pyramid_request.session.peek_flash('success')
+
+    def test_if_validation_succeeds_it_calls_to_xhr_response(self,
+                                                             matchers,
+                                                             pyramid_request,
+                                                             to_xhr_response):
+        form_ = conftest.form_validating_to('anything')
+
+        form.handle_form_submission(pyramid_request,
+                                    form_,
+                                    mock.Mock(),
+                                    mock.Mock())
+
+        to_xhr_response.assert_called_once_with(
+            pyramid_request,
+            matchers.redirect_302_to(pyramid_request.url),
+            form_)
+
+    def test_if_validation_succeeds_it_returns_to_xhr_response(self,
+                                                               pyramid_request,
+                                                               to_xhr_response):
+        result = form.handle_form_submission(pyramid_request,
+                                             conftest.form_validating_to('anything'),
+                                             mock.Mock(),
+                                             mock.Mock())
+
+        assert result == to_xhr_response.return_value
+
+    @pytest.fixture
+    def to_xhr_response(self, patch):
+        return patch('h.form.to_xhr_response')
